@@ -20,7 +20,7 @@
         "exclude": ["/docs",  "/docs/build"]
     },
     "require": {
-        "php": ">=7.4"
+        "php": ">=8.0"
     },
     "autoload": {
         "classmap": [
