@@ -26,10 +26,11 @@ class Calculator extends AbstractCalculator implements NnAndNzCounterIF{
     protected $nz = [];
 
     /**
+     * @param MajorPropertiesService $mp Object of major properties  
+     * @param int $maxLevel The maximum (group) level 
      * Initialize all levels with 0 values
      */
-    public function __construct(MajorProperties $mp, int $maxLevel) {
-        parent::__construct($mp, $maxLevel);
+    public function __construct(protected MajorProperties $mp, public int $maxLevel) {
         $this->total = $this->nz = $this->nn = array_fill(0, $maxLevel + 1, 0);
     }
 
