@@ -14,14 +14,19 @@ declare(strict_types=1);
 namespace gpoehl\phpReport\getter;
 
 /**
- * Gets value from a data row using a callable
+ * Gets value from an array item
  */
-class GetFromCallable extends BaseGetter{
-   
+class GetArrayItemForJoin extends BaseGetter
+{
+
     /**
-     * @see BaseGetter::getValue
+     * @see getValue()
+     * @return Array item declared in $source. 
+     * Warning for non existing items will be suppressed. Required for joins.
+     * There it will end with a no data action. 
      */
     public function getValue($row, $rowKey = null) {
-        return ($this->source)($row, $rowKey, ...$this->params);
+        return @$row[$this->source];
     }
+
 }
