@@ -14,14 +14,16 @@ declare(strict_types=1);
 namespace gpoehl\phpReport\getter;
 
 /**
- * Gets value from an object property where the data row is the object.
+ * Gets value from an static property. Not related to a data row.
+ * Class can be a class name or an object.
  */
-class GetFromObjectProperty extends BaseGetter{
+class GetStaticProperty extends BaseGetter{
    
     /**
      * @see BaseGetter::getValue
      */
-    public function getValue($row, $rowKey =null) {
-        return $row->{$this->source};
+    public function getValue($row, $rowKey = null) {
+        // Note the ${$name[1]} syntax!
+        return $this->source[0]::${$this->source[1]};
     }
 }
