@@ -11,7 +11,7 @@ use gpoehl\phpReport\Group;
 class GroupTest extends TestCase {
 
     public function testInstantiateGroup() {
-        $group = new Group('groupA', 2, 1, 'testSource');
+        $group = new Group('groupA', 2, 1, 'testSource', []);
         $this->assertSame('groupA', $group->name);
         $this->assertSame(2, $group->level);
         $this->assertSame(1, $group->dimID);
@@ -20,11 +20,11 @@ class GroupTest extends TestCase {
     
     public function testInvalidLevel() {
         $this->expectException(InvalidArgumentException::class);
-        $group = new Group('groupA', 0, 1, 'testSource');
+        $group = new Group('groupA', 0, 1, 'testSource', []);
     }
      public function testInvalidDimID() {
         $this->expectException(InvalidArgumentException::class);
-        $group = new Group('groupA', 1, -3, 'testSource');
+        $group = new Group('groupA', 1, -3, 'testSource', []);
     }
 
 }
