@@ -24,12 +24,12 @@ class GroupsTest extends TestCase {
         $this->assertSame(1, count($groups->groupLevel));
         $this->assertSame(0, $groups->groupLevel['total']);
 
-        $groups->addGroup(new Group('groupA', 1, 0, 1));
+        $groups->addGroup(new Group('groupA', 1, 0, 1, []));
         $this->assertSame(1, count($groups->items));
         $this->assertSame('groupA', $groups->items[1]->name);
         $this->assertSame(1, $groups->groupLevel['groupA']);
 
-        $groups->addGroup(new Group('groupB', 2, 0, 1));
+        $groups->addGroup(new Group('groupB', 2, 0, 1, []));
         $this->assertSame(2, count($groups->items));
         $this->assertSame('groupB', $groups->items[2]->name);
         $this->assertSame(2, $groups->groupLevel['groupB']);
@@ -37,21 +37,15 @@ class GroupsTest extends TestCase {
 
     public function testAddGroupNameTwiceWillFail() {
         $groups = $this->stack;
-        $groups->addGroup(new Group('groupA', 1, 0, 1));
+        $groups->addGroup(new Group('groupA', 1, 0, 1, []));
         $this->expectException(InvalidArgumentException::class);
-        
-        
-//        $this->expectExceptionMessage();
-        
-        
-        
-        $groups->addGroup(new Group('groupA', 2, 0, 1));
+        $groups->addGroup(new Group('groupA', 2, 0, 1, []));
     }
 
     public function testFirstGroupLevelIsNotOne() {
         $groups = $this->stack;
         $this->expectException(InvalidArgumentException::class);
-        $groups->addGroup(new Group('groupA', 2, 0, 1));
+        $groups->addGroup(new Group('groupA', 2, 0, 1, []));
     }
 
     /**
@@ -59,16 +53,16 @@ class GroupsTest extends TestCase {
      */
     public function testAddGroupInvalidLevel($group) {
         $groups = $this->stack;
-        $groups->addGroup(new Group('group1', 1, 0, 1));
-        $groups->addGroup(new Group('group2', 2, 0, 1));
+        $groups->addGroup(new Group('group1', 1, 0, 1, []));
+        $groups->addGroup(new Group('group2', 2, 0, 1, []));
         $this->expectException(InvalidArgumentException::class);
         $groups->addGroup($group);
     }
 
     public function groupProvider() {
         return [
-            ['Level is less' => new Group('group3', 1, 1, 1)],
-            ['Level is too high' => new Group('group5', 5, 1, 1)],
+            ['Level is less' => new Group('group3', 1, 1, 1,[])],
+            ['Level is too high' => new Group('group5', 5, 1, 1,[])],
         ];
     }
 
